+++ src/score.rs
@@ -1,13 +1,13 @@
 use crate::player::Player;
 
 #[derive(Debug, Copy, Clone)]
-pub struct Score(pub u32, pub u32);
+pub struct Score(pub u8, pub u8);
 impl Score {
     pub fn new() -> Score {
         Score(0, 0)
     }
 
-    pub fn of(self, player: Player) -> u32 {
+    pub fn of(self, player: Player) -> u8 {
         match player {
             Player::One => self.0,
             Player::Two => self.1,
